+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/model/policy/Option.kt
@@ -1,0 +1,6 @@
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.policy
+
+data class Option(
+  val value: String,
+  val conditional: Conditional?,
+)
