+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/model/policy/PassInput.kt
@@ -1,0 +1,8 @@
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.policy
+
+data class PassInput(
+  val type: String,
+  val label: String,
+  val name: String,
+  val includeBefore: String?,
+)
