+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/model/policy/StandardConditions.kt
@@ -1,0 +1,10 @@
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.policy
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class StandardConditions(
+  @JsonProperty("AP")
+  val standardConditionsAp: List<StandardConditionAp>,
+  @JsonProperty("PSS")
+  val standardConditionsPsses: List<StandardConditionPss>,
+)
